SNS streamline GPIO interface
