+++ lib/io/gpio.hpp
@@ -1,35 +1,40 @@
 #pragma once
 
-#include "gpio_reader.hpp"
-#include "gpio_writer.hpp"
-
 #include <cstdint>
 #include <optional>
 
 #include <core/logger.hpp>
+#include <core/types.hpp>
 
 namespace hyped::io {
 
 /**
- * An abstract GPIO interface. This is to be used in all places where GPIO access is
- * required. It is the callees responsibility to provide a correct implementation
- * such as `Gpio` below.
+ * An abstract interface to read from a GPIO pin. This is to be used whenever read access
+ * to GPIO is required.
+ */
+class IGpioReader {
+ public:
+  virtual core::DigitalSignal read() = 0;
+};
+
+/**
+ * An abstract interface to write to a GPIO pin. This is to be used whenever write access
+ * to GPIO is required.
+ */
+class IGpioWriter {
+ public:
+  virtual void write(const core::DigitalSignal state) = 0;
+};
+
+/**
+ * An abstract GPIO interface. This is to be used in all places where it is necessary to
+ * initiate GPIO access. It is the callees responsibility to provide a correct implementation
+ * such as `Gpio` in `real_gpio.hpp`.
  */
 class IGpio {
  public:
   virtual std::optional<IGpioReader> getReader(const std::uint8_t pin) = 0;
   virtual std::optional<IGpioWriter> getWriter(const std::uint8_t pin) = 0;
-};
-
-class Gpio {
- public:
-  Gpio(hyped::core::ILogger &log);
-
-  std::optional<GpioReader> getReader(const std::uint8_t pin);
-  std::optional<GpioWriter> getWriter(const std::uint8_t pin);
-
- private:
-  hyped::core::ILogger &log_;
 };
 
 }  // namespace hyped::io
