--- lib/io/gpio_reader.cpp
@@ -1,11 +1,0 @@
-#include "gpio_reader.hpp"
-
-namespace hyped::io {
-
-core::DigitalSignal GpioReader::read()
-{
-  // TODO: implement
-  throw -1;
-}
-
-}  // namespace hyped::io
