--- lib/io/gpio_writer.cpp
@@ -1,10 +1,0 @@
-#include "gpio_writer.hpp"
-
-namespace hyped::io {
-
-void GpioWriter::write(const core::DigitalSignal state)
-{
-  // TODO: implement
-}
-
-}  // namespace hyped::io
